// src/app/model/IOrder.ts
export interface IOrder {
    id: string;
    tracking_Number: string;
    status: string;
    origin: string;
    destination: string;
    address: string;
    name: string;
    comment: string;
    company: string;
    initiated: string;
    dispached: string;
    customs: string;
    delivered: string;
    closed: string;
    weight: string;
}